multiple_ports: Let Each Peer Have Multiple Ports
- Use ports queries and responses instead of socket queries. Send
a query with a message that contains all the updated ports to the
server. Make the server send the streams ip and the retrieved ports
to the server manager that responds with the ip map from peer ids
to ports and the port map with ips to a vector of ports. Make the
server send these back to the peer so it can set it's fields.
Update decoding functions accordingly.
- Create a get_connection() function in src/network/peer.rs to
try to connect to an ip with a certain number of ports
- Create a listen() function in src/network/server.rs to handle
listening to connections on multiple ports
- Adjust new transaction message code to allow for ips with
multiple ports